refactoring modals into single component & balance check fix
