+++ public/css/balance-modal.css
@@ -1,8 +1,6 @@
 /* Balance Modal */
 .balance-modal {
   font-family: "Avenir", Helvetica, Arial, sans-serif;
-  width: 730px;
-  padding: 15px;
 }
 .balance-modal h3 {
   margin: 0 0 15px;
