+++ public/css/instructions-modal.css
@@ -1,8 +1,6 @@
 /* Instructions modal */
 .instructions-modal {
   font-family: "Avenir", Helvetica, Arial, sans-serif;
-  width: 755px;
-  min-height: 418px;
 }
 .instructions-modal h3 {
   margin: 0 0 25px;
