+++ public/css/modal.css
@@ -34,7 +34,7 @@
   text-transform: uppercase;
   font-size: 17px;
   padding: 15px;
-  width: 700px;
+  width: 755px;
   visibility: hidden;
   opacity: 0;
   transition: all ease 0.3s;
